Make DispatchQueue drain synchronously on main
This is a fix for the same issue detailed in this CL:
https://android-review.googlesource.com/c/platform/frameworks/support/+/1145596/

That CL changed to use Dispatchers.Main.immediate, but Main and
Main.immediate have identical dispatch functions, the difference lies in
the isDispatchNeeded function.

This changes so DispatchQueue checks isDispatchNeeded, and if not
needed, drains the queue synchronously. This ensures that coroutines
started with launchWhenCreated happen synchronously as part of an
onCreate (and not dispatched to the next frame). A test has been added
for this case as well to illustrate the desired behavior.

Test: ./gradlew lifecycle:lifecycle-ktx:test
Change-Id: I178af70da3752d128de4f3b48f13a81bca8ba1a4
(cherry picked from commit 12ea2590020307f4baf599c3dd63438558dff996)
(cherry picked from commit a04409d113ea54461cf8f18261d223ddd19876b5)