Handle boxing in field vs its getter/setter
This CL fixes a bug in room where we were not using the right type
adapters when a field's type does not exactly match its getter/setter
type.
For instance, if field is Integer and getter returns `int`, the code
we generate would do a null check on that getter, causing a compilation
failure.

The core issue was that `isAssignable` always returns true between
boxed types and we were always inferring the type from field (because
of the assumption that getter/setter has the same type).
Now the PojoProcessor updates the reader/writer adapter if the
getter/setter type does not match the field's type. We also put a
warning in this case as it might cause unexpected values in the
database.

Bug:142780009
Test: BoxedPrimitivesTest, TableEntityProcessorTest
Change-Id: Ic1592d66a01000b32696d0c69b60536f8f5550ef