Fix testOnVideoSizeChanged in media2-widget
It modifies testOnVideoSizeChanged not to depend on the current media
item to fix flakiness as it is not guaranteed that the current media
item is changed when onVideoSizeChanged is called.

It also adds checking the aspect ratio of the surface view is equal to
the ratio of the video size. Without this check, the test was just
testing the player not VideoView.

Bug: 140270775
Test: ./gradlew :media2:media2-widget:connectedCheck
Change-Id: I53357fff4b42258f816b3110ace7c3b8fb98a24d