Add biometric api dependency on lifecycle-viewmodel
Fixes a bunch of errors (red squiggles) currently being reported for the
biometric library in Android Studio. The current stable version of
lifecycle-viewmodel doesn't have the missing interface that's causing
the issue, so just use the project version for now. We can later switch
to 2.2.0 for the release branch, or just leave it out entirely.

Test: Open Android Studio, sync Gradle, no more red squiggles
Test: Sanity-check existing functionality in the biometric demo app
Change-Id: I4ad40b62beb89ee62c9d7d8e07b9386a4b132da9