Serialization schema type hierarchy
This CL adds the basic type hierarchy for serialization-schema as a
hierachy of sealed classes. Sealed classes enables clients to lean on
the Kotlin compiler to verify the exhaustivity of when statements on
types, but in their current implementation, all the subtypes must be in
the same file as the base type.

The classes defined in this CL are sufficient to describe a
serialization schema, but additional helper classes will be added in a
subsequent CL to make it easy to build schemas with types that
reference each other.

Test: ./gradlew :serialization:serialization-schema:test
Bug: 143380082
Change-Id: Ida57e867e231d1eb694110e44685783a9927647f