Fix deep link parsing only first character
When combining path and query param placeholders in a deep link, only
the first character or the path parameter gets parsed.

This is because the regex that is used to match the final path
parameter before the question takes the minimum amount of characters to
satify the matcher. What we would actually like to do is have all of the
character up to the question mark be matched.

This change adjust the regex to look for either the end of string (in
the case that all of the parameters are optional) or for the literal
question mark and any characters after it.

Test: Added NavDeepLinkTest
BUG: 144554689
Change-Id: I5615d7bb3ca2257b0d5b42351d28b836b3cf2b2c