Handle the case of an action without an icon.
Below API 23, we extract the resource ID of the icon for the action
when converting it from ActionCompat to Action in the method
getActionFromActionCompat. However, it is possible that no icon was
defined by passing a 0 resource ID when creating the action. In this
case, the IconCompat will be null and unconditionally trying to access
its resource ID leads to a NullPointerException.

This commit handles that case correctly by checking for null and for
cases in which the IconCompat was not created from a resource ID and
use the 0 resource ID instead.

This specific behavior is covered by tests in NotificationCompatTest
already, which are currently broken for API level 17-22, but adding a
few additional tests to specific focus on this seems like a good idea.

This issues was introduced in:
  https://android-review.googlesource.com/1147238
when moving off some deprecated methods, but without including the
null check.

Bug: b/144040975
Test: ./gradlew core:core:connectedAndroidTest
Change-Id: Ib25801e2bcca567059eb1043ac503708e540b3b4