Moved PreviewSurfaceProviders to camera-testing
PreviewSurfaceProviders is moved to camera-testing and renamed to
SurfaceTextureProvider since it is simply a convenience method to
produce a SurfaceTexture. Usages of it might be removed in the future
from any tests that do not strictly require a SurfaceTexture.

Applications should either produce their own SurfaceTexture for GL
rendering or use PreviewView for simple Preview display.

Bug: b/144869228
Test: ./gradlew bOS
Change-Id: Id0fa1a88e3236bdadfa760ec39813451bee548d2