DayNight rewrite v2.5
We've now moved to using createConfigurationContext(),
instead of applyOverrideConfiguration(). Certain older versions
of WebView have an issue when an app uses
applyOverrideConfiguration(), but they seem to work fine
using createConfigurationContext() /shruggie.

To use createConfigurationContext(), I've had to refactor a
lot of updateNightMode() code to also have another entry
point which returns a Context for attachBaseContext().

Test: ./gradlew appcompat:connectedCheck
BUG: 141351441

Change-Id: I7a5b59071f82b79820c6df695bfa622d727f2030