Add benchmarks for RadioGroup.
I was investigating some tests locally and had to write these anyway so
it makes sense to check them in also.

There is nothing particulary shocking, locally I have:
18ms on recomposition for this benchmark
(just for illustration: 4ms recomposition for the checkbox one).

I think the reason is that the RadioGroup uses a pattern where variable
is observed from forEach. That leads to recomposition of all the radios
on model change. I could work-around it but if this is pattern we
encourage I would prefer to use it.

Bug: b/136262330
Test: Added
Change-Id: I9d380816668a6878709100ff6e0424a04d448355