Change read observations to allow observing non-ComponentNodes.
RepaintBoundariesNodes will be removed in the future and replaced
by a modifier or similar. DrawNodes will also be removed now that
DrawModifiers exist.  Model read observations cannot be tied to
ComponentNodes.

This CL changes NodeStagesModelObserver (renamed to ModelObserver)
to use nested observation. The observations are tied to a
commit callback and a target object that will be passed to the
commit callback. Observation is done on a lambda block and
can likewise be paused within a lambda block rather than using
stages.

Test: AndroidLayoutDrawTest and demo
Change-Id: Ia4d93e4dd5d554a2e7dce5b94f4a4535967f31e1