Initial support for inserting separators between paginated items
This will allow apps to e.g. create letter headers in a contacts list,
or date separators in a list of events.

Note: separators at the beginning / end of the entire list are not yet
supported. To support these, we need PageEvents to convey whether
they're at the end or not, which will come later.

Bug: 111130543
Test: ./gradlew paging:paging-common:test

Change-Id: Icd6072265be3ab97ef63ab49e91953a3174d5879