Adjust the default fragment transition presets
The current fragment transition presets did not take into consideration
that fragments rarely have a non-opaque background. This change adjusts
the fragment presets with specifications specifically for fragments.

The open and close transitions now have a duration of 300 instead of 400
and the fade transition has a duration of 150. The open transition fades
the exiting view to an alpha of 0 instead of 0.4.

Test: Tested in muliple applications
BUG: 145468417
Change-Id: I6e1be474d0bb218e4a519bb09ea5bc1fc1757dcf