Fix NPE in FragmentManager moveToState when using Robolectric
If an exiting fragment has an animation with a duration of 0, instead of
posting the animation start and end like a normal device, robolectric
instantly ends the animation. This causes the fragment's container to be
nulled out immediately, and as a result a NPE is thrown when the
fragment's view is being removed from the container.

This ensures that if the animation is ended immediately, we have a
reference to the container so we can still properly remove the view.

Test: added robolectric test
BUG: 145770616
Change-Id: I7956d558f524268be509539e146bf432ac560abf