Revert to using FixedSizeSurfaceTexture for preview.
This CL reverts to using FixedSizeSurfaceTexture with both PreviewView's TextureViewImplementation and within CameraView. This is a temporary fix for the distorted preview seen on both PreviewViewFragment and CameraViewFragment after onPause/onResume (b/146048690).
This fix is set to be released (if stable) as part of camera-view Alpha09.

I manually tested this on a couple of Level 3 and full devices (wasn't able to find a full device). @scottnien it would be great if you can test it on the pixel 2 you mentioned in the bug.

Bug: N/A
Test: Manual testing The CameraView and PreviewView fragments on the camerax view test app
Change-Id: I169c8d115ee1e48f3ff579fd02aef4fc96183b49