Telling Gradle to initially request all of the memory that it will ever request
(only when DIST_DIR is set)

to rule out the possibility that the Gradle daemon might be unable to acquire more memory from the host system in the middle of the build and then crashing with an out-of-memory error.

Also, in cases where Gradle does eventually use as much memory as allowed, this should slightly improve performance.

Bug: 146217083
Test: DIST_DIR=../../out/dist ./gradlew --no-daemon projects # and see that it doesn't take any more time to run with this change vs without
Test: ./gradlew --no-daemon projects
Test: DIST_DIR=../../out/dist ./gradlew --no-daemon --info 2>&1 | grep -i Xmx | head -n 1 | grep Xms
Test: ./gradlew --no-daemon --info 2>&1 | grep -i Xmx | head -n 1 | grep Xms # and see there are no matches

Change-Id: Id6c0c45c375a3f1a836af0d473f9146f3f99d103