Fix ImageCapture MaxQuality test timeout
Root cause:
The FakeRepeatingUseCase which used in  MaxQuality test might close
the ImageReader immediately after the use case is unbound.
And the ImageReader in the FakeRepeatingUseCase  might be closed before
the capture session is closed. It might cause the exception:
"java.lang.IllegalArgumentException: Surface had no valid native Surface."
and the camera will callback the onError().

Solution:
Register the setOnSurfaceDetachedListener and only close the ImageReader
after the surface was detached.

Test: ./gradlew bOS &&
./gradlew camera:camera-core:connectedAndroidTest &&
./gradlew camera:camera-camera2:connectedAndroidTest &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:camera-core:test &&
./gradlew camera:camera-camera2:test &&
./gradlew camera:camera-extensions:test &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:integration-tests:camera-testapp-extensions:connectedAndroidTest

Change-Id: I4b599277a812866d09881fe30391b7749964be6f