Fix "ImageReaderContext is not initialized" crash which cause test failure and app issue.
In api level 23 and below,  calling imageReader.acquireNextImage(or acquireLatestImage) when
ImageReader is closed results in a RuntimeException  "ImageReaderContext is not initialized"

In api level 24+ , it'll return a null Image instead.

To fix the issue and make it consistent across all api levels,  make it return null when
the RuntimeException happens.

Bug: 147027763
Test: AndroidImageReaderProxyTest
Change-Id: I1c1147244bab3b23f84423e37ab2898dde21c78b