Avoid try/catching exceptions in FragmentPagerAdapters
When FragmentPagerAdapter or FragmentStatePagerAdapter
runs finishUpdate() in response to a measure/layout
pass, it runs commitNowAllowingStateLoss() as all
normal devices don't do measure/layout while processing
pending FragmentTransactions, but Robolectric will. In
order to ensure that Robolectric continued to work,
b/137201343 provided behavior to fallback to
commitAllowingStateLoss() when processing pending
FragmentTransactions. It used a try/catch for
IllegalStateExceptions, which means valid exceptions
could be ignored.

This change drops any transactions that attempt to be
committed while the FragmentManager is already executing
transactions.

Test: FragmentPagerAdapterRoboTest
Fixes: 147359544

Change-Id: I21ebbfe67d349980003b6df77704f005e693e664