RecyclerView MergeAdapter
This CL introduces a new Adapter type for RecyclerView that merges other
adapters.
It isolates all item types of adapters (unless otherwise given in the
config), forwards all callbacks to the necessary adapters (like create,
binding, onrecycle etc).

It is not optimized for having hundreds of adapters as it seems very
unlikely to be a real use case.

It does not currently support stable ids which can be added later.
The case for using stable ids is not much as there is already a
ListAdapter that can calculate the diff for the developer.

Bug: 147833132
Test: MergeAdapterTest
Change-Id: I1d89cf665dbf28c8532617aa60cbb1ffb7154545