Ensure <include-dynamic> graphs are restored
The included graphs from installed modules
normally replace the <include-dynamic> graph
instances when you navigate() to them, but this
breaks when you restore the state of the
NavController since it requires that the graph have
all of the destinations on the back stack.

By keeping track of all of the DynamicIncludeNavGraph
instances we create, we can ensure that when
onRestoreState does occur, we replace any of them
that are from an installed module to avoid
"IllegalStateException: unknown destination during
restore" issues.

Test: Ran the sample app in the bug
BUG: 148316337
Change-Id: I98b57ef375be49d1f602260340eb35c69600af87