Fix bug where JPEG file with XMP is incorrectly saved
Currently, even if a JPEG file has an XMP data stored separately
from the EXIF data, it saves it as an EXIF tag, and when
ExifInterface#saveAttributes is called, the XMP data is stored
inside the EXIF data. This can cause a problem since the maximum
bytes that an EXIF data can hold inside a JPEG file is 65536 bytes,
and having the XMP data inside the EXIF data can make the EXIF data
exceed this limit.

This CL solves this issue by separately saving the EXIF data if it
existed separately in the original file.

Bug: 149179921
Test: manually with file attached in bug
Change-Id: I3be3df83cbd12df7fae8c578a87bad8dca6e5781