Refactors AppBars to use Box + modifiers
This cl also adds touch target size to AppBarIcon, so we can simplify
the padding around AppBar - the only padding that currently needs to
exist is 4.dp at the start and end, but we should clarify this in the
future. This has the intentional side effect of letting custom slots
take up the full space of the AppBar (barring the 4.dp mentioned above),
instead of enforcing unecessary padding.

Bug: b/143752110
Test: AppBarTest
Change-Id: I779d336deb234d92ee01fad9524d1371cbca45a0