BUGFIX: failed ImageCapture test
The test introduced in the previous CL aosp/1233439 is failing on some devices:

https://sponge.corp.google.com/target?id=daf22744-d15e-4dd1-a537-c62f5bcb7f0c&target=//third_party/java_src/android_libs/camerax/mobileharness/unit/core:ImageSaverTest_on_sm-g610f&searchFor=&show=FAILED&sortBy=STATUS

It seems that MediaStore behaves differently on different devices/OS. On local test device, ContentResolver.insert() returns null for invalid data column. On test lab devices, it returns a URI but FNFE later. Updated the test to be less specific.

Bug: 149177060
Test: manual test and ./gradlew bOS
Change-Id: I1a6e5cbaab6ba6b6d822896e6f6de60165b80409