Fixed nested RepaintBoundaries issues on 29
When we draw RepaintBoundaries on 29 and then we want to update some property of which doesn't trigger the content reinvalidation(opacity, elevation) we don't redraw the parent RepaintBoundary. But previously we had code like this "if (node.opacity > 0f) renderNode.draw(canvas)". Which means we didn't draw the render node at all, so once it's opacity will be changed from 0 we will not redraw as the parent rendernode invalidation is not triggered. We should draw the children rendernode even if it is currently transparent. Exactly similar issue with not drawing the shadow when we switch from 0 to any non-0 is fixed as well.

Bug: 149481241
Test: manually and covered with new tests
Change-Id: I15de6f9ac8579862eaf84e1636895f29d77f9a89