Replace DrawerLayout with Openable in NavigationUI
Rather than rely directly on the concrete
subclass of DrawerLayout, use the Openable
interface instead.

As DrawerLayout implements the new Openable interface,
this is source compatible with the previous
API surface. The variables on the AppBarConfiguration.kt
extensions are specifically kept as drawerLayout to
ensure that named parameters require no changes.

Test: ran testapp, tests still pass
BUG: 129030452
Relnote: "Instead of relying on the concrete `DrawerLayout` class,
`AppBarConfiguration` now uses the `Openable` interface (which
`DrawerLayout` implements)."
Change-Id: Id3aba402d17e0203d40b1a5c9558c32b60fddc11