Make animation clock controllable in tests
In tests, the ambient animation clock that is provided at the root of
the composition is now swapped out for a TestAnimationClock that can be
paused, advanced manually, resumed, and queried if it is idle. An idle
clock is one that has no animations subscribed to it.

The TestAnimationClock is used by default in tests that have an
AndroidComposeTestRule. If you want the test clock without using that
rule, use an AnimationClockTestRule.

Relnote: "Tests using AndroidComposeTestRule now provide an animation
clock at the root of the composition that allows it to be paused,
resumed and advanced manually."

Bug: 145133230
Bug: 150286087
Test: AnimationSynchronizationTest

Change-Id: Id54c51482f554cdb512e4eb53c78930408778f94