Refactors Tabs to have slots for text and icon
This allows us to simplify how we handle the configuration for
text and/or icon, and also allows us to use Icons directly in samples,
making it possible to fully copy and paste samples.

Also exposes a base 'Tab' overload that is not opinionated about layout,
and cleans up the FancyTab sample to use this.

Replaces Containers with Boxes

Stops using `sp` for bottom-baseline-alignment, these should be in DP to
avoid the labels from being at the wrong vertical alignment when the
font scale changes.

Bug: b/145903416
Bug: b/150285641
Test: TabTest
Relnote: Refactored Tab APIs to accept `text` and `icon` lambdas,
instead of String and Image. See updated samples for usage information.

Change-Id: Ia057e449977c4115ae6d564fdad341562030351e