Deprecate Fragment's onActivityCreated
The original purpose of the onActivityCreated() callback was to allow
fragment logic to be tied to the creation of its activity. We should no
longer be encouraging this coupling.

We should promote passing in external dependencies as construction
parameters using FragmentFactory. Code touching the fragment's view
should be done in onViewCreated() and other initialization code should
be in onCreate(). To receive a callback specifically when the activity's
onCreate() is complete, a LifeCycleObserver should be registered on the
activity's Lifecycle in onAttach(), and removed once the onCreate()
callback is received.

Test: ./gradlew checkApi
BUG: 144309266
Change-Id: Iba62d27bc6b87aa5048315a4cb90e5dcdf550114