move diff util test to kotlin
This CL moves the diff util test to kotlin. Just a service
to the future myself not to get annoyed by the verbose code
in this class. There is more cleanup to do, like moving to
truth but i thought it is better to keep this CL just a
language change.

Bug: n/a
Test: test passes
Change-Id: I930d1dde73505b3bd2c5e0034eba812475195627