Fix Separators generics
Test: ./gradlew paging:p-c:test

Separators had the generic constraint that (R : T), when in reality we
wanted (T : R). That can only be done in an extension method, since
it's invalid to constrain a local generic class type.

We additionally (manually) add a static PagingData#insertSeparators
method for Java callers, as having it in a separate class is not
discoverable.

Added sample code in both Kotlin and Java to prove UiModel pattern can
work with new extension and static method.

Change-Id: I4d87485b5d87ec4186b1f8cdfff012d0d0b047f6