Remove enabled param from FABs
According to Material specs Disabled states cannot be inherited by floating action buttons. If the action represented in the FAB is unavailable, the FAB should not appear in the composition. To enforce it via the API we remove enabled param from the FAB components making it clear that instead of disabling you should just don't add FAB.

Fixes: 150847904
Test: manually, only api change
Change-Id: Ifbb345d2d5b8f9d24e3e0e264583eb4e2c722d90