Introduce LoadedFontFamily which takes Typeface
A Typeface is a loaded object of the FontFamily.
To be able to use Typeface to the styling interface, wrap Typeface with
FontFamily. During loading or resolving typeface, loader don't do
anything and just use the internal Typeface.

This CL introduces LoadedFontFamily but nobody can create it since at
this moment since nobody can get Typeface from FontFamily.

Bug: 143715852
Test: ./gradlew test
Test: TreeHugger
Change-Id: Ib1ec194731dec86a3f6d164b735e917fb0051be5