[Semantics] SCN -> PassthroughLayout(modifier) and remove caches
This is the first big Semantics modifier patch.  This changes the
implementation details of semantics heavily but without touching
API or tests at all yet (except 3 tests that are sensitive to
PassThroughLayout).  I plan to follow up by patches to change the
Semantics API later in April.

A) This updates the implementation details of semantics to internally
use a modifier, without changing the API surface at all yet.  The only
supported use of the new modifier is
"PassThroughLayout + single Semantics modifier".
Any other usage, for now, is not expected to work.

Modifier chaining is not supported yet and it still
uses a single-child-composable-style approach.  In particular,
semantics nodes still apply to the closest "non-semantics" layout node
underneath them, and "collapsing" semantics works the same as before.

B) To simplify the transition to a new design in the future, this patch
also removes all cached state -- it now always constructs derived
values at the last minute from the original source of truth.
Specifically, I deleted caches involved in
1) collapsing (topNodeOfConfig/_children/_unmergedChildren)
2) merging (_mergedChildren, mergedParent, mergedConfig)
3) attaching/detaching (SemanticsOwner MutableMap<>s), a concept
which has been removed in favor of always doing depth-first-search.

This means in turn that no cache invalidation/markNeedsUpdate is needed,
which I noticed was the main source of correctness bugs.

Fixes: 147606016
Fixes: 147606282
Fixes: 150777826
Test: existing tests
Change-Id: I64879097cb05cf08491adec4b599619fd6a8e3de