Expose changed and dependent sets for AMD
This refactors the AffectedModuleDetector so that instead
of only returning the affectedModules, it also keeps the
projects sorted into the changed/dependent sets.

The change also exposes those sets so that other tasks
can use them.
Also a minor bugfix for cobuilt projects: Cobuilts previously
only worked for the changed target. Now it will also work for dependent.

Bug: 151744308
Test: Added unit tests and existing tests pass
Change-Id: I564215b8f9a446aa2b75c9386d8a399f4b883287