Fix Autofill in Demo App
aosp/1206889 introduced this bug when it changed ambients to a more efficient implementation. We missed out initializing the Autofill ambient. This CL restores the initialization for the autofill ambient.

This regression was not caught by the tests because the test was using owner.autofill instead of using the autofill ambient.

Fixes: 153280962
Test: Verified that autofill works in the Demo app. Updated the test to use the autofill ambient instead of owner.autofill
Change-Id: I41959b106fca027ed4b436b1555c04d4e2a2c350