Root LayoutNode to behave as a Stack
This means we will measure the direct children with loosen constraints (minWidth and minHeight = 0) instead of behaving as if Modifier.fillMaxSize() was already applied as it proven to be not predictable and making writing any sample projects a bit too complex. Almost always you have to start with something like
setContent {
    Stack {
        Button()
    }
}
instead of just directly emit Button without a Stack. When you wrote
setContent {
    Button()
}
You probably didn't mean I wanted to fill the whole screen with a huge Button. Put a default sized Button into a left top corner of the screen seems like a more reasonable default

Relnote: We changed how we measure the first layout you put inside activity.setContent { } block. Previously it was forced to fill the whole activity screen, and now it behaves as if you put your layout inside a Stack: it can be smaller than a screen and will be positioned in the top left screen. If you want the old behaviour you can apply Modifier.fillMaxSize() for your layout.
Bug: 153453714
Test: manually, new tests
Change-Id: Ie88a713401b8d1db32acf421f2612121bd1d23f6